Updates example application
