+++ geocoding/example/pubspec.yaml
@@ -9,6 +9,8 @@ environment:
   sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  baseflow_plugin_template: ^2.1.2
+  
   flutter:
     sdk: flutter
