+++ geocoding/pubspec.yaml
@@ -1,6 +1,6 @@
 name: geocoding
 description: A Flutter Geocoding plugin which provides easy geocoding and reverse-geocoding features.
-version: 2.1.0
+version: 2.1.1
 repository: https://github.com/baseflow/flutter-geocoding/tree/main/geocoding
 issue_tracker: https://github.com/Baseflow/flutter-geocoding/issues
